requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
numpy==1.22.3
Pillow==9.0.1
requests==2.27.1
torch==1.10.2+cpu
torchvision==0.11.3+cpu
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
